fix cow type

+++ src/main/java/com/capstone/dfms/requests/FeedMealDetailRequest.java
@@ -8,13 +8,15 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
 public class FeedMealDetailRequest {
 
-    private Long quantity;
+    private BigDecimal quantity;
 
     private Long itemId;
